+++ src/com/samuli/GameRoundPair.java
@@ -2,10 +2,10 @@ package com.samuli;
 
 public class GameRoundPair {
     public Game game;
-    public double error;
+    public int error;
     public Round round;
 
-    public GameRoundPair(Game _game, Round _round, double _error) {
+    public GameRoundPair(Game _game, Round _round, int _error) {
         game = _game;
         error = _error;
         round = _round;
